[ADD] mrp_project
================
MRP Project Link
================

It links projects and tasks to *manufacturing orders* (MO) and
*work orders* (WO).

Usage
=====

In a manufacturing order (MO), you can select a project to be attached to it.
If none is selected, when the MO is confirmed, a project is automatically
created.

When the MO starts, a task is created and assigned to the order.