[FIX] mrp_production_real_cost: Update costs when operation time lines change
Up to now, you have to rely only on play/pause/stop controls, and the time passed
between them for having real costs reflected, but there are manufacturing processes
that reflects their times not directly, but afterwards introducing the summaries.
This commit allows to reflect the changes on costs when you edit up times.