[14.0][FIX] mrp_multi_level: error with 3 step delivery.
If the destination moves do not include a MO, the mrp move end
up without a name that raises an error later. This can happen
with 3 setep delivery but also with in more cases in wich there is
a chain of stock moves.

We ensure that the variable `order_number` always have a value.