[14.0][MIG] mrp_kit_drop_ship
Without this module, if a product is being procured and has an existing Bill of
Materials of type Kit, the components of the product are going to be
procured, instead of the main product.

This module introduces the capability to ignore the existence of the kit
during procurement, when the procurement is associated with a dropship. That
is, when the product is going to be sourced from a third party vendor to the
customer.

Upon confirmation of the purchase order the system will still explode the
product into components within the drop ship picking.