mrp_lot_number_propagation: Clean outdated finished serial number
When a manufacturing order is split through the produce all wizard,
the serial numbers of the finished product are created and written on each
backorder MO, before move lines are unlinked from the original MO and
created again on the backorders.

As long as nothing changes on the propagating component, we are fine.
But in case the component is scrapped or unreserved, and another component
serial number was assigned, we had a discrepancy between the propagating
lot number from the component and the lot number of the finished product.

Since this module creates and assigns the finished serial number during
the call to button_mark_done, we could have wiped the serial number as soon as
the move line propagating its serial number was unassigned. But in that case
Odoo would still have created a stock.lot record for each backorder that
would have been correct in most cases, just to be recreated afterwards.

Instead, we will clean the outdated serial number of the finished
product only in case a propagating move with a different serial number
is assigned to the manufacturing order, as well as protecting against
any discrepancy during the when calling button_mark_done.